requirements.txt
langchain>=1.2.15
python-dotenv>=1.2.2
langchain-openai>=1.1.12
ddgs>=9.13.0
streamlit>=1.56.0
flask>=3.1.3
flask_cors>=6.0.2
